+++ functions/api/routes/form/formDetailAnswerPinPUT.js
@@ -3,7 +3,7 @@ const util = require('../../../lib/util');
 const statusCode = require('../../../constants/statusCode');
 const responseMessage = require('../../../constants/responseMessage');
 const db = require('../../../db/db');
-const slackAPI = require('../../../middlewares/slackAPI');
+const slackAPI = require('../../../lib/slackAPI');
 const { answerDB } = require('../../../db');
 
 module.exports = async (req, res) => {
