+++ functions/api/routes/team/teamFeedbackCreatePOST.js
@@ -3,7 +3,7 @@ const util = require('../../../lib/util');
 const statusCode = require('../../../constants/statusCode');
 const responseMessage = require('../../../constants/responseMessage');
 const db = require('../../../db/db');
-const slackAPI = require('../../../middlewares/slackAPI');
+const slackAPI = require('../../../lib/slackAPI');
 const { feedbackDB, linkFeedbacKeywordDB, keywordDB, userDB } = require('../../../db');
 const dayjs = require('dayjs');
