+++ functions/lib/crypto.js
