+++ functions/lib/resizeImage.js
