+++ functions/lib/slackAPI.js
